--- zulu/src/main/java/us/ihmc/zulu/parameters/controller/ZuluICPControllerParameters.java
@@ -1,7 +1,0 @@
-package us.ihmc.zulu.parameters.controller;
-
-import us.ihmc.commonWalkingControlModules.capturePoint.controller.ICPControllerParameters;
-
-public class ZuluICPControllerParameters extends ICPControllerParameters
-{
-}
