--- zulu/src/main/java/us/ihmc/zulu/parameters/controller/ZuluStepAdjustmentParameters.java
@@ -1,28 +1,0 @@
-package us.ihmc.zulu.parameters.controller;
-
-import us.ihmc.commonWalkingControlModules.capturePoint.stepAdjustment.StepAdjustmentParameters;
-
-public class ZuluStepAdjustmentParameters extends StepAdjustmentParameters
-{
-   private static final boolean useStepAdjustment = true;
-
-   /** {@inheritDoc} */
-   @Override
-   public boolean allowStepAdjustment()
-   {
-      return useStepAdjustment;
-   }
-
-   /** {@inheritDoc} */
-   @Override
-   public double getAdjustmentDeadband()
-   {
-      return 0.005;
-   }
-
-   @Override
-   public double getMinimumTimeForStepAdjustment()
-   {
-      return -0.2;
-   }
-}
